+++ inst/examples/recogito/tests/shinytest/tagging.R
@@ -1,0 +1,4 @@
+app <- ShinyDriver$new("../../")
+app$snapshotInit("tagging")
+
+app$snapshot()